@@ -24,14 +24,14 @@ var errSetRawInvalidValue = fmt.Errorf("invalid value type")
 // Message is the data structure that is handled by transforms and interpreted by
 // conditions.
 //
-// Data in each message may be JSON text or binary data:
+// Data in each message can be accessed and modified as JSON text or binary data:
 //   - JSON text is accessed using the GetValue, SetValue, and DeleteValue methods.
 //   - Binary data is accessed using the Data and SetData methods.
 //
-// Metadata is a second data field that is meant to store information about the message,
-// but can be used for any purpose. For JSON text, metadata is accessed using the
-// GetValue, SetValue, and DeleteValue methods with a key prefixed with "meta ". Binary
-// metadata is accessed using the Metadata and SetMetadata methods.
+// Metadata is an additional data field that is meant to store information about the
+// message, but can be used for any purpose. For JSON text, metadata is accessed using
+// the GetValue, SetValue, and DeleteValue methods with a key prefixed with "meta" (e.g.
+// "meta foo"). Binary metadata is accessed using the Metadata and SetMetadata methods.
 //
 // Messages can also be configured as "control messages." Control messages are used for flow
 // control in Substation functions and applications, but can be used for any purpose depending
@@ -46,10 +46,12 @@ type Message struct {
 	ctrl bool
 }
 
+// String returns the message data as a string.
 func (m *Message) String() string {
 	return string(m.data)
 }
 
+// New returns a new Message.
 func New(opts ...func(*Message)) *Message {
 	msg := &Message{}
 	for _, o := range opts {
@@ -59,16 +61,21 @@ func New(opts ...func(*Message)) *Message {
 	return msg
 }
 
-func AsControl() func(*Message) {
-	return func(m *Message) {
-		m.ctrl = true
-	}
+// AsControl sets the message as a control message.
+func (m *Message) AsControl() *Message {
+	m.data = nil
+	m.meta = nil
+
+	m.ctrl = true
+	return m
 }
 
+// IsControl returns true if the message is a control message.
 func (m *Message) IsControl() bool {
 	return m.ctrl
 }
 
+// Data returns the message data.
 func (m *Message) Data() []byte {
 	if m.ctrl {
 		return nil
@@ -77,6 +84,7 @@ func (m *Message) Data() []byte {
 	return m.data
 }
 
+// SetData sets the message data.
 func (m *Message) SetData(data []byte) *Message {
 	if m.ctrl {
 		return m
@@ -86,6 +94,7 @@ func (m *Message) SetData(data []byte) *Message {
 	return m
 }
 
+// Metadata returns the message metadata.
 func (m *Message) Metadata() []byte {
 	if m.ctrl {
 		return nil
@@ -94,6 +103,7 @@ func (m *Message) Metadata() []byte {
 	return m.meta
 }
 
+// SetMetadata sets the message metadata.
 func (m *Message) SetMetadata(metadata []byte) *Message {
 	if m.ctrl {
 		return m
@@ -103,6 +113,14 @@ func (m *Message) SetMetadata(metadata []byte) *Message {
 	return m
 }
 
+// GetValue returns a value from the message data or metadata.
+//
+// If the key is prefixed with "meta" (e.g. "meta foo"), then
+// the value is retrieved from the metadata field, otherwise it
+// is retrieved from the data field.
+//
+// This only works with JSON text. If the message data or metadata
+// is not JSON text, then an empty value is returned.
 func (m *Message) GetValue(key string) Value {
 	if strings.HasPrefix(key, metaKey) {
 		key = strings.TrimPrefix(key, metaKey)
@@ -117,6 +135,14 @@ func (m *Message) GetValue(key string) Value {
 	return Value{gjson: v}
 }
 
+// SetValue sets a value in the message data or metadata.
+//
+// If the key is prefixed with "meta" (e.g. "meta foo"), then
+// the value is placed into the metadata field, otherwise it
+// is placed into the data field.
+//
+// This only works with JSON text. If the message data or metadata
+// is not JSON text, then this method does nothing.
 func (m *Message) SetValue(key string, value interface{}) error {
 	if strings.HasPrefix(key, metaKey) {
 		key = strings.TrimPrefix(key, metaKey)
@@ -141,6 +167,14 @@ func (m *Message) SetValue(key string, value interface{}) error {
 	return nil
 }
 
+// DeleteValue deletes a value in the message data or metadata.
+//
+// If the key is prefixed with "meta" (e.g. "meta foo"), then
+// the value is removed from the metadata field, otherwise it
+// is removed from the data field.
+//
+// This only works with JSON text. If the message data or metadata
+// is not JSON text, then this method does nothing.
 func (m *Message) DeleteValue(key string) error {
 	if strings.HasPrefix(key, metaKey) {
 		key = strings.TrimPrefix(key, metaKey)
@@ -170,34 +204,42 @@ type Value struct {
 	gjson gjson.Result
 }
 
-func (v Value) Value() interface{} {
+// Value returns the value as an interface{}.
+func (v Value) Value() any {
 	return v.gjson.Value()
 }
 
+// String returns the value as a string.
 func (v Value) String() string {
 	return v.gjson.String()
 }
 
+// Bytes returns the value as a byte slice.
 func (v Value) Bytes() []byte {
 	return []byte(v.gjson.String())
 }
 
+// Int returns the value as an int64.
 func (v Value) Int() int64 {
 	return v.gjson.Int()
 }
 
+// Uint returns the value as a uint64.
 func (v Value) Uint() uint64 {
 	return v.gjson.Uint()
 }
 
+// Float returns the value as a float64.
 func (v Value) Float() float64 {
 	return v.gjson.Float()
 }
 
+// Bool returns the value as a bool.
 func (v Value) Bool() bool {
 	return v.gjson.Bool()
 }
 
+// Array returns the value as a slice of Value.
 func (v Value) Array() []Value {
 	var values []Value
 	for _, r := range v.gjson.Array() {
@@ -207,10 +249,12 @@ func (v Value) Array() []Value {
 	return values
 }
 
+// IsArray returns true if the value is an array.
 func (v Value) IsArray() bool {
 	return v.gjson.IsArray()
 }
 
+// Map returns the value as a map of string to Value.
 func (v Value) Map() map[string]Value {
 	values := make(map[string]Value)
 	for k, r := range v.gjson.Map() {
@@ -220,6 +264,7 @@ func (v Value) Map() map[string]Value {
 	return values
 }
 
+// Exists returns true if the value exists.
 func (v Value) Exists() bool {
 	return v.gjson.Exists()
 }
