@@ -75,7 +75,7 @@ func dynamodbHandler(ctx context.Context, event events.DynamoDBEvent) error {
 
 		// CTRL messages flush the pipeline. This must be done
 		// after all messages have been processed.
-		ctrl := message.New(message.AsControl())
+		ctrl := message.New().AsControl()
 		if _, err := sub.Transform(tfCtx, ctrl); err != nil {
 			return err
 		}
