@@ -71,7 +71,7 @@ func snsHandler(ctx context.Context, event events.SNSEvent) error {
 
 		// CTRL messages flush the pipeline. This must be done
 		// after all messages have been processed.
-		ctrl := message.New(message.AsControl())
+		ctrl := message.New().AsControl()
 		if _, err := sub.Transform(ctx, ctrl); err != nil {
 			return err
 		}
