@@ -157,7 +157,7 @@ func main() {
 
 		// ctrl messages flush the pipeline. This must be done
 		// after all messages have been processed.
-		ctrl := message.New(message.AsControl())
+		ctrl := message.New().AsControl()
 		if _, err := sub.Transform(ctx, ctrl); err != nil {
 			return err
 		}
