@@ -70,9 +70,10 @@ func newAWSSecretsManager(_ context.Context, cfg config.Config) (*awsSecretsMana
 	}
 
 	c.client.Setup(aws.Config{
-		Region:     conf.AWS.Region,
-		RoleARN:    conf.AWS.RoleARN,
-		MaxRetries: conf.Retry.Count,
+		Region:          conf.AWS.Region,
+		RoleARN:         conf.AWS.RoleARN,
+		MaxRetries:      conf.Retry.Count,
+		RetryableErrors: conf.Retry.ErrorMessages,
 	})
 
 	return c, nil
