@@ -8,7 +8,6 @@ import (
 
 	"github.com/brexhq/substation/config"
 	"github.com/brexhq/substation/message"
-	mess "github.com/brexhq/substation/message"
 )
 
 var _ Transformer = &aggregateFromArray{}
@@ -102,7 +101,7 @@ func TestAggregateFromArray(t *testing.T) {
 
 			// aggregateFromArray relies on an interrupt message to flush the buffer,
 			// so it's always added and then removed from the output.
-			ctrl := message.New(mess.AsControl())
+			ctrl := message.New().AsControl()
 			messages = append(messages, ctrl)
 
 			tf, err := newAggregateFromArray(ctx, test.cfg)
