@@ -8,7 +8,6 @@ import (
 
 	"github.com/brexhq/substation/config"
 	"github.com/brexhq/substation/message"
-	mess "github.com/brexhq/substation/message"
 )
 
 var _ Transformer = &aggregateToArray{}
@@ -122,7 +121,7 @@ func TestAggregateToArray(t *testing.T) {
 
 			// aggregateToArray relies on an interrupt message to flush the buffer,
 			// so it's always added and then removed from the output.
-			ctrl := message.New(mess.AsControl())
+			ctrl := message.New().AsControl()
 			messages = append(messages, ctrl)
 
 			tf, err := newAggregateToArray(ctx, test.cfg)
