@@ -6,7 +6,6 @@ import (
 
 	"github.com/brexhq/substation/config"
 	"github.com/brexhq/substation/message"
-	mess "github.com/brexhq/substation/message"
 	"golang.org/x/exp/slices"
 )
 
@@ -108,7 +107,7 @@ func TestAggregateToString(t *testing.T) {
 
 			// aggregateToString relies on an interrupt message to flush the buffer,
 			// so it's always added and then removed from the output.
-			ctrl := message.New(mess.AsControl())
+			ctrl := message.New().AsControl()
 			messages = append(messages, ctrl)
 
 			tf, err := newAggregateToString(ctx, test.cfg)
