@@ -87,13 +87,29 @@ type metaSwitch struct {
 
 func (tf *metaSwitch) Transform(ctx context.Context, msg *message.Message) ([]*message.Message, error) {
 	if msg.IsControl() {
+		var messages []*message.Message
 		for _, c := range tf.conditional {
-			if _, err := c.tf.Transform(ctx, msg); err != nil {
-				return nil, fmt.Errorf("transform: meta_switch: %v", err)
+			res, err := c.tf.Transform(ctx, msg)
+			if err != nil {
+				return nil, fmt.Errorf("transform: meta_pipeline: %v", err)
 			}
+
+			messages = append(messages, res...)
 		}
 
-		return []*message.Message{msg}, nil
+		// This is required to deduplicate the control messages that
+		// were sent to the conditional transforms.
+		var msgs []*message.Message
+		for _, m := range messages {
+			if m.IsControl() {
+				continue
+			}
+
+			msgs = append(msgs, m)
+		}
+
+		msgs = append(msgs, msg)
+		return msgs, nil
 	}
 
 	for _, c := range tf.conditional {
