@@ -74,9 +74,10 @@ func newSendAWSDynamoDB(_ context.Context, cfg config.Config) (*sendAWSDynamoDB,
 	}
 
 	tf.client.Setup(aws.Config{
-		Region:     conf.AWS.Region,
-		RoleARN:    conf.AWS.RoleARN,
-		MaxRetries: conf.Retry.Count,
+		Region:          conf.AWS.Region,
+		RoleARN:         conf.AWS.RoleARN,
+		MaxRetries:      conf.Retry.Count,
+		RetryableErrors: conf.Retry.ErrorMessages,
 	})
 
 	agg, err := aggregate.New(aggregate.Config{
